Extract Material UI config to a separate directory
