+++ packages/api-docs-builder/projects/materialUi/projectSettings.ts
@@ -1,0 +1,15 @@
+import path from 'path';
+import { ProjectSettings } from 'buildApi';
+import findApiPages from '../../utils/findApiPages';
+import getMaterialUiComponentInfo from './getMaterialUiComponentInfo';
+
+const projectSettings: ProjectSettings = {
+  output: {
+    apiManifestPath: path.join(process.cwd(), 'docs/data/material/pagesApi.js'),
+  },
+  typeScriptProjects: ['material', 'lab'],
+  getApiPages: () => findApiPages('docs/pages/material-ui/api'),
+  getComponentInfo: getMaterialUiComponentInfo,
+};
+
+export default projectSettings;
